+++ backend/intents/foods.bad.json
@@ -1,0 +1,8 @@
+{
+    "questions": [
+        "What's your least favorite food?"
+    ],
+    "answers": [
+        "Any vegetables are gross.  In my home country, all we eat is bread."
+    ]
+}
