+++ backend/intents/foods.good.json
@@ -1,0 +1,8 @@
+{
+    "questions": [
+        "What's your favorite food?"
+    ],
+    "answers": [
+        "PIZZA"
+    ]
+}
