+++ backend/intents/greetings.bye.json
@@ -1,0 +1,17 @@
+{
+    "questions": [
+        "goodbye",
+        "bye take care",
+        "see you later",
+        "bye for now",
+        "i must go",
+        "adios"
+    ],
+    "answers": [
+        "see you soon!",
+        "Till next time",
+        "bye bye",
+        "have a great day",
+        "huh?"
+    ]
+}
