+++ backend/package.json
@@ -8,6 +8,7 @@
   },
   "author": "",
   "dependencies": {
+    "express": "^4.17.2",
     "fastify": "^3.27.1",
     "node-nlp": "^4.22.18"
   }
