--- server/intents/food.json
@@ -1,1 +1,0 @@
-{"questions":["What's your favorite food?","What's your least favorite food?"],"answers":["PIZZA","Any vegetables are gross.  In my home country, all we eat is bread."]}
