--- server/intents/thing.json
@@ -1,1 +1,0 @@
-{"questions":["adios"],"answers":["huh?"]}
